Moved Endpoint impl into type definition file. (See [comment](https://github.com/diesel-rs/diesel/pull/4169#discussion_r1719596303))
