+++ examples/postgres/custom_arrays/Cargo.toml
@@ -8,5 +8,5 @@ publish = false
 diesel = { path = "../../../diesel", version = "2.2.0", default-features = false, features = ["postgres", "r2d2"] }
 diesel_migrations = { path = "../../../diesel_migrations", version = "2.2.0", default-features = false, features = ["postgres"] }
 
-
-dotenv = "0.15.0"
+[dev-dependencies]
+dotenvy = "0.15"
